Extract state for GeneAutocompletion into GeneAutocompleteBase
